enh: add tracking library
