+++ front/lib/tracking/amplitude/browser/ampli.json
