+++ front/lib/tracking/amplitude/config.ts
