+++ front/lib/tracking/amplitude/node/ampli.json
