+++ front/pages/api/user/index.ts
@@ -7,10 +7,10 @@ import * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { trackUserMemberships } from "@app/lib/amplitude/node";
 import { updateUserFullName } from "@app/lib/api/user";
 import { getSession } from "@app/lib/auth";
 import { getUserFromSession } from "@app/lib/iam/session";
+import { ServerSideTracking } from "@app/lib/tracking/node";
 import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
@@ -50,7 +50,7 @@ async function handler(
 
   switch (req.method) {
     case "GET":
-      trackUserMemberships(user).catch((err) => {
+      ServerSideTracking.trackUserMemberships({ user }).catch((err) => {
         logger.error(
           { err: err, userId: user.sId },
           "Failed to track user memberships"
