+++ front/pages/index.tsx
@@ -48,10 +48,10 @@ import Particles from "@app/components/home/particles";
 import ScrollingHeader from "@app/components/home/scrollingHeader";
 import { PricePlans } from "@app/components/PlansTables";
 import { SubscriptionContactUsDrawer } from "@app/components/SubscriptionContactUsDrawer";
-import { trackPageView } from "@app/lib/amplitude/browser";
 import { getSession } from "@app/lib/auth";
 import { getUserFromSession } from "@app/lib/iam/session";
 import { makeGetServerSidePropsRequirementsWrapper } from "@app/lib/iam/session";
+import { ClientSideTracking } from "@app/lib/tracking/browser";
 import { classNames } from "@app/lib/utils";
 
 const { GA_TRACKING_ID = "" } = process.env;
@@ -122,7 +122,7 @@ export default function Home({
   }, []);
 
   useEffect(() => {
-    trackPageView({
+    ClientSideTracking.trackPageView({
       pathname: router.pathname,
     });
   }, [router.pathname]);
